+++ constructorio-client/src/main/java/io/constructor/client/models/AutocompleteResponse.java
@@ -16,6 +16,9 @@ public class AutocompleteResponse {
     @SerializedName("result_id")
     private String resultId;
 
+    @SerializedName("request")
+    private Map<String, Object> metadata;
+
     /**
      * @return the resultId
      */
