+++ constructorio-client/src/main/java/io/constructor/client/models/RecommendationsResponse.java
@@ -1,5 +1,7 @@
 package io.constructor.client.models;
 
+import java.util.Map;
+
 import com.google.gson.annotations.SerializedName;
 
 /**
@@ -12,7 +14,10 @@ public class RecommendationsResponse {
 
     @SerializedName("response")
     private RecommendationsResponseInner response;
-    
+
+    @SerializedName("request")
+    private Map<String, Object> metadata;
+
     /**
      * @return the resultId
      */
