+++ constructorio-client/src/test/java/io/constructor/client/ConstructorIORecommendationsTest.java
@@ -17,18 +17,6 @@ public class ConstructorIORecommendationsTest {
     public ExpectedException thrown = ExpectedException.none();
 
     @Test
-    public void createRecommendationsResponseShouldReturnAResult() throws Exception {
-        String string = Utils.getTestResource("response.recommendations.home_page_1.json");
-        RecommendationsResponse response = ConstructorIO.createRecommendationsResponse(string);
-
-        assertEquals("pod display name exists", response.getResponse().getPod().getDisplayName(), "You may also like");
-        assertEquals("pod display id exists", response.getResponse().getPod().getId(), "item_page_1");
-        assertEquals("recommendation results exist", response.getResponse().getResults().size(), 10);
-        assertEquals("total number of results", (int)response.getResponse().getTotalNumberOfResults(), 10);
-        assertTrue("recommendation result id exists", response.getResultId() != null);
-    }
-
-    @Test
     public void getRecommendationsShouldReturnAResultWithSingleItemId() throws Exception {
         ConstructorIO constructor = new ConstructorIO("", "ZqXaOfXuBWD4s3XzCI1q", true, null);
         UserInfo userInfo = new UserInfo(3, "c62a-2a09-faie");
