+++ constructorio-client/src/test/java/io/constructor/client/NaturalLanguageSearchRequestTest.java
@@ -12,6 +12,7 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 public class NaturalLanguageSearchRequestTest {
+
     @Rule
     public ExpectedException thrown = ExpectedException.none();
