+++ constructorio-client/src/test/java/io/constructor/client/RecommendationsRequestTest.java
@@ -9,6 +9,7 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 public class RecommendationsRequestTest {
+
     @Rule
     public ExpectedException thrown = ExpectedException.none();
